Remote CLI wrapper scripts
As part of the providing remote CLI capabilities, we provide
docker images for both platform and application, each
containing clients preinstalled.

To make remote clients easier to use, we also provide a set
of configuration and wrapper scripts. These scripts take as
input an RC file downloaded from Horizon (platform or
application horizon) and generate a script with the proper
authentication variables and aliases.

For more details, a README file is also included.

These scripts also work on Windows, by using cygwin to
interpret the bash scripts

Note: these files are not supposed to be included in the
ISO or be built as part of the normal build. These files
are to be distributed separately as a tarball.

Change-Id: I00f6358b8b984150fc11e9087afba83063cb8bb3
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>
Closes-Bug: 1834667
Closes-Bug: 1834668