Support custom image tags for remote CLI images
As part of creating a distributable tarball or remote CLI
wrapper script, we must also be able to package an exact
version of docker image files to ensure compatibility
between the remote CLI tarball and the distributed ISO and
helm chart.

This change writes the desired image tag in a separate file
that can then be overwritten by the packaging script as
needed.

Change-Id: Iab99486f0a8e697e3cec861ba9a45d9794c21f3a
Closes-bug: 1840133
Depends-on: I156172a6ed208d6fcf9bb8f37182daea73b2856c
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>