Fix remote dcmanager access
Dcmanager needs another environment variable set in order to properly
work with the remote setup.

The variable is OS_ENDPOINT_TYPE and must be set to "publicURL" for
any kind of remote access. Unless otherwise specified, dcmanager
defaults to "internalURL".

The issue only affect dcmanager, as other platform commands (system
and fm) work without this variable set, but also don't seem to be
affected in any way when this variable is configured to "publicURL".

Change-Id: Ic70de40ff169d07800f9ee42f744b12768dc70ac
Closes-bug: 1869258
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>