Add volume map to keep the repositories.yml
This solution is to keep the data from the helm
config. The helm repo commands are managed locally
and that is the reason that the command helm repo
add works, the file doesn't exist but this command
creates that, but when it runs helm repo list the
file repositories.yml doesn't exist anymore because
the container execution didn't keep that.

Attention: In the solution it is mapping the default
path /root/.config/helm/ and /root/.cache/helm/
inside the container, if for some reason the it is
changed passing a flag --repository-config or
--repository-cache this new path must be changed in
the client_wrapper.sh file, otherwise it will come
back to the same behavior.

There is no reason to change this default path, it is
inside the container of the remote cli, but we have
to keep it in mind, for some possible future issues.

Closes-bug: 1925802
Signed-off-by: Rafael Jordão Jardim <RafaelJordao.Jardim@windriver.com>
Change-Id: I3180513b76c75221b6143a6da4d45a600e477662
(cherry picked from commit 8a7f010ce3e79f16644cdbf5b0b19268dfe1a88f)