Change the entrypoint of the image to always use bash
To fix the kubectl version in the commit that it depends on, the
CUSTOMIZATION variable of the file
stx-platformclients.stable_docker_image had to be used.

But when we are building a new image with CUSTOMIZATION, the
build-stx-images.sh script put a CMD ["bash"] in the image, making the
old way to call commands in stx-platformclients to no longer work.

This commit solves this issue, by forcing the entrypoint to always
use bash and making all commands run as with "bash -c <command>"

Test Plan:
- Tested this change with new and old stx-platformclients versions.
  It work with the image that has the kubectl fix and the older ones.

Partial-Bug: #1955427
Depends-on: https://review.opendev.org/c/starlingx/upstream/+/822356

Signed-off-by: Thiago Miranda <ThiagoOliveira.Miranda@windriver.com>
Change-Id: Ieae227daee875bc4b4b42509cbcb78350360d02c